[PoC] Asset Fetcher: implement pipelines
This is an experiment with a more capable and flexible asset/fetcher.
It describes assets that can be chained in a pipeline.  The common
use case is downloading an archive, and then extracting a file from
it.

As a "caching" mechanism, the pipeline runner will start from the last
position, and works its way back.

In theory, this should allow for asset setup in very flexibly ways,
including:
 * Creating snapshots
 * Running "guestfish" like commands on OS images (verifier would
   check if action was indeed performed)

A practical and simple example is at `examples/asset_pipeline.py`.

Signed-off-by: Cleber Rosa <crosa@redhat.com>