nrunner: add status service and server implementation
One of the task's responsibility is to provide status updates on its
execution.  The actual methods, protocols and channels that a task can
use to send those status updates are not strictly determined.

This initial implementation uses a simple TCP socket based
client/server.  The task execution should not hang or be delayed by
attempts to post the status updates, so the asyncio framework is used
to allow for concurrent execution and delivery of status updates.

Signed-off-by: Cleber Rosa <crosa@redhat.com>