avocado.utils.memory: allow the module to be loaded on non Linux
Even though the individual features of this module are Linux only, it
doesn't mean that the module shouldn't be capable of load under other
platforms.

Now, if users attempt to fetch information, and the implementation is
Linux only, undefined behavior will follow.

In the future, a better solution is needed, such as a new namespace
for features that are really Linux only, or some form of extensible
implementation for other platforms.

Signed-off-by: Cleber Rosa <crosa@redhat.com>