core: New package 'future' with the first module: 'settings'
The settings module is a new and experimental configuration handler and
has intention to handle both: command line and configuration files. This
is part of our discussion "Configuration by Convention", and it is a
preparation to remove all the default values from different parts of
code. This package itself should not affect the current Avocado usage.

Reference: https://avocado-framework.readthedocs.io/en/latest/blueprints/BP001.html
Signed-off-by: Beraldo Leal <bleal@redhat.com>