future/settings: Migrate replay to use the new future/settings API
As part of the effort to remove default values scattered around the
code, this change only migrates the replay plugin options to use the new
module. This will associate the following mapping between args and
config options: 1) --replay, run.replay.job_id; 2) --replay-ignore,
run.replay.ignore; 3) --replay-test-status, run.replay.test_status and
4) --replay-resume, run.replay.resume. It is important to mention that
even if a command-line argument has the intention to run only once and
change this value from a configuration file is not needed, with this
model we need to have a "namespace" (section + key) anyway.

Signed-off-by: Beraldo Leal <bleal@redhat.com>