Spawner: add interface for waiting on tasks
In my observation, under some circunstances, the use of the futures
to monitor the process has been unstable.  Then, in some use cases,
such as when explicitly monitoring tasks and waiting for them, it
makes sense to have a more direct interface to do so.

This introduces an interface and implementantions that do just that.

Signed-off-by: Cleber Rosa <crosa@redhat.com>