[NOT FOR MERGE] N(ext) Runner Job plugin: migrate to TaskStateMachine
This shows how the integration of the TaskStateMachine into this
runner implementation will, more or less, look like.

The feature set achieved here is that tests are running in parallel,
and the tasks are being managed by the TaskStateMachine and Workers.
And, the integration with the job is largely done, given that all
results are generated.

A feature that is *missing* here is that ResultEvent plugins are not
yet updated while the state machine is running and while the status
messages are being received.  That is coming shortly.

Signed-off-by: Cleber Rosa <crosa@redhat.com>