Introduce Status Server implementation
A server that listens on the network for status messages and uses a
StatusRepo to store them.

This version is also different than the one at
avocado/core/status_server.py, as it support UNIX domain sockets,
which will be handy when creating a communication location between job
and tests in the job result directory itself.

Signed-off-by: Cleber Rosa <crosa@redhat.com>