Introduce requirements caching module
The tasks of checking for requirements, one of the first phases of
the N(ext) Runner scheduler, is one that can be pretty expensive.
It makes sense to have utilities that will record the result of
requirements verification.

Implementation wise, this builds on top of SQLITE because:

 * It's built into Python
 * It allows multiple connections to the same database file
   for reading operations
 * It allows proper locking when multiple attempts to write
   to database are performed

With regards to the database schema, a partial normalization was
chosen because it allows for straightforward use of the database, that
is, inserts are performed with the real data and queries show
meaningful data, instead of requiring a view or more complicated query.

Signed-off-by: Cleber Rosa <crosa@redhat.com>