nrun: make use of task requirement check
Even the the nrun plugin is about to be removed, this makes sure we
touch on the code for checking the requirements (and to a minor
extent, add a "cheap" feature).

One limitation here inherit to the nrun plugin is that the checks for
all tasks are done at a single place, and no task actually start
running before the requirements of all actually finish.

So this change also helps to make it clear how the characteristics
discussed on BP003, an on an upcoming implementation, are not bound by
those limitations.

Signed-off-by: Cleber Rosa <crosa@redhat.com>