Make Task identifier unique
There's no need to manually set the identifier of a Task, and it
shouldn't be set if it's not going to be an unique one per job.

Note: there should of course be a deduplication of requirements, but
since it's not one yet, this would prevent multiple requirements from
conflicting at the task identifier level.

Signed-off-by: Cleber Rosa <crosa@redhat.com>