selftests/unit/test_utils_network.py: only test protocols if address available
Tests were failing due to the fact that, even when there are no IPv6
addresses available, and attempt to bind to addresses (such as
"localhost") on IPv6 was attempted.  But, the same could be the
case for IPv4.

This now only activates those checks when there are other valid
addresses.

Signed-off-by: Cleber Rosa <crosa@redhat.com>