Skip Decorators: fix a bug when no condition for skip exists
When no condition for skip exists, the test should *not* be skipped,
that is, it should run.  But, the current decorator was not executing
the decorated function.

This was being masqueraded because the default behavior for tests that
do nothing is to pass, and by not running the decorated function (the
test) it would effectively pass.  Now the tests make sure that the
test method gets executed (and does a cancel when running it).

Signed-off-by: Cleber Rosa <crosa@redhat.com>