Task and Messages: add debugging information on Tasks' progress
This adds debugging information to classes that interact with Tasks
(namely TaskStateMachine, Worker and StatusRepo).  The goal is
to be able to debug the progress of a Task more easily.

The loggers are named after their location in the Avocado module
structure, which I believe are also meaningful and easy to remember.
For instance, one can do:

 $ avocado --show=avocado.core run --test-runner=nrunner /bin/true

To see all internal (core) Avocado messages, or:

 $ avocado --show=avocado.core.task run --test-runner=nrunner /bin/true

To see only the task related ones.

Fixes: https://github.com/avocado-framework/avocado/issues/4547
Signed-off-by: Cleber Rosa <crosa@redhat.com>