avocado-safeloader-find-python-unittest: introduce unittest compatible order
Python's unittest module uses a, well, interesting order for the tests
it discovers.  The order does *not* necessarily match the order found
in the source code file.

Avocado's safeloader, on the other hand, does respect the order of the
tests in the source code file.

This introduces a mode that returns the findings in the same order as
Python's unittest, so that everything else about their behavior can be
compared.

Signed-off-by: Cleber Rosa <crosa@redhat.com>