safeloader: introduce module for keeping track of imported symbols
The "safeloader.module.PythonModule" users a dictionary with
filesystem path-like value for keeping track of imported modules.
This is not the most effective format for that information, given that
it assumes the current module location as the start of the path.  If
the import is referring to an importable symbol that lives in another
directory, the filesystem path-like will not follow and be aware of
that location.

This introduces a proper module that expands the amount of information
on imported symbols so that a better informed decision can be made
when trying to find the actual modules files (for further safe
parsing).

Signed-off-by: Cleber Rosa <crosa@redhat.com>