safeloader: support both relative and absolute imports
This fixes an issue that was manifested in imports from
"selftests.utils" from Avocado's tests in "selftests/unit",
"selftests/functional", etc.

There are tests in "selftests/functional/test_loader.py" which cover
the safeloader code with regards to relative imports, so it's safe to
standardize on a single style (absolute import, previously borken) on
the actual selftests.

Special mention to the selftests/safeloader.sh test that now checks
every single file that should have Python unittests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>