Runnable: allow configuration of its identifier
So far, the user visible identification of a runnable has always been
the value that is given in its "uri" attribute.  But, there are
situations in which other characteristics of the runnable contain more
information about what it does.

This allows a configuration based on the name of the attributes, as a
simple mechanism of per-runnable configuration.

One possibility from here, not attempted because of its increased
complexity (and lack of current use case) is to allow a formatting
string that can receive the runnable attributes for substitution.

Signed-off-by: Cleber Rosa <crosa@redhat.com>