Assets pre-tests plugin: enable support for nrunner
When using nrunner, requirements (including assets) can be added to
docstrings and they will be resolved as tasks that tests dependend on.

But, there's no reason why implicit assets can not be fetched before
the job if the assets pre-tests plugin is enabled when using nrunner.

At the worst case scenario, this can be used to complete the removal
of the legacy runner without losing test coverage.

Signed-off-by: Cleber Rosa <crosa@redhat.com>