Drop wrapper support on legacy runner
Avocado's legacy runner has support for the --wrapper feature.

They depend on "hacks" (and hooks) on the avocado.utils.process, and
rely on one test process running at a time. The very similar --gdb
feature was removed in the past because of similar reasons.

So that the mission of removing the legacy runner can be accomplished,
let's remove this feature.

An issue exists to track the re-implementation (or not) of such a
feature for the nrunner.

Reference: https://github.com/avocado-framework/avocado/issues/5353
Signed-off-by: Cleber Rosa <crosa@redhat.com>