Introduce "ansible-module" runner/dependency support
This new plugin allows dependencies to be expressed and resolved using
any available ansible module.

For instance, instead of asking test writers to write code to say,
create user accounts that will be used during a test, a developer can
simply use ansible's "user" module.

The same goes for instance, when a test needs a service installed or
up and running: instead of writing all that code, the test writer can
describe that dependency using the "service" ansible module.

Signed-off-by: Cleber Rosa <crosa@redhat.com>