avocado.utils.linux_modules: protect when /proc/modules is not available
Instead of raising a FileNotFoundError when attempting to get a list
of all modules by reading /proc/modules, if it's not available, let's
return an empty list (and log that fact).

The main reason is this is used indirectly by higher level
functionality, such as decorators to skip tests in environments where
there is no Linux module functionality, so it needs to behave a bit
nicer.

Signed-off-by: Cleber Rosa <crosa@redhat.com>