avocado/core/task/statemachine.py: guarantee match between encodings
The encoding used on all encode() calls by default, is one hard coded
in the Python implementation.  For cpython, this is:

   const char *
   PyUnicode_GetDefaultEncoding(void)
   {
       return "utf-8";
   }

But, the fact that we are advertising the encoding used (in the
"encoding" key of the message) implies that we should not rely on
anything external (including the Python implementation) for that
information.  It should be perfectly possible to have the producer
side sending data with a different encoding than the one on the
consumer side, as long as the enconding used is known.

Signed-off-by: Cleber Rosa <crosa@redhat.com>