reduce IO load from slower aexpect's process pipes
The bulk of runner messages sent to the status server via the network
are also included in the runner's stdout which ends up hogging the
entire aexpect process. Since such stdout is not used anywhere and
our debug.log is already collected via the messages straight to the
status server, dropping it already helps the test perform on par with
equivalent tests from other spawners.

TODO: Test on normal test set too.