Avocado Instrumented Test: introduce a multiplier for the timeout
This introduces a "timeout factor", basically another parameter that
Avocado Instrumented tests will respect that will be used to alter the
actual timeout.

The reason for this feature is the fact that tests are coded with
timeouts, but, their actual needed timeouts will depend on the
environment they are running on.  It's not OK to expect the test code
to be modified when the environment changes significantly for an
execution of the tests (say on faster or slower CPUs or networks).

The implementation is done in a way that adds information to the
avocado.Test class.  The only possible breakage with regards to
compatibility is if custom user's test classes already define an
"actual_timeout" member, which is easily fixable by a simple rename.

Signed-off-by: Cleber Rosa <crosa@redhat.com>