CI: enable the container image directive
A number of GitHub Actions run as part of the CI checks are not
effectively setting the container image that should be used.  This
effectively means that instead of covering a number of different
platforms, it's re-running the same checks on the base OS.

This applies the directive that effective makes GitHub action set up
the container where the job will be run.

Signed-off-by: Cleber Rosa <crosa@redhat.com>