CI: fix installation of Python and setuptools
The "version" and "egg" tasks use official container images for a
number of operating systems that may use either DNF or APT as a
package manager, and may or may not have Python or setuptools.

The existing checks and installation is broken, and was not detected
because they were always running on the base GitHub actions image.

Signed-off-by: Cleber Rosa <crosa@redhat.com>