CI: fix installation of Python and setuptools for different OS images
The "version" and "egg" tasks use official container images for a
number of operating systems that may use either DNF or APT as a
package manager, and may or may not have Python or setuptools.

The existing checks and installation is broken, and was not detected
because they were always running on the base GitHub actions image.

A number of GitHub Actions run as part of the CI checks are not
effectively setting the container image that should be used.  This
effectively means that instead of covering a number of different
platforms, it's re-running the same checks on the base OS.

This applies the directive that effective makes GitHub action set up
the container where the job will be run.

Signed-off-by: Cleber Rosa <crosa@redhat.com>