Podman spawner: tag images created by the spawner due to requirements
The podman spawner will create images with requirements so that they
can be reused by other tests.  But those images won't have any
meaningful tags associated with them.  That is, users doing a "podman
images" command will only see "<UNKNOWN>".

With this change, a somewhat meaningful description is attempted to be
used as tag.

For instance, when running:

 $ avocado run --spawner=podman \
   -- examples/tests/passtest_with_dependency.py

Which has:

   :avocado: dependency={"type": "package", "name": "hello"}

As a docstring directive, under a Fedora 38 machine, the
image will be tagged:

   localhost/avocado_generated_fedora_38_package_hello

The prefix is configurable and should also allow users to filter/clean
images created/cached by Avocado.

Signed-off-by: Cleber Rosa <crosa@redhat.com>